requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.2.0
django-extensions==2.0.6
django-fe-core==0.3.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-nested-routers==0.90.2
factory-boy==2.10.0
Faker==0.8.12
idna==2.6
mysqlclient==1.3.12
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
six==1.11.0
text-unidecode==1.2
urllib3==1.22
